[ANDROSDK-1987] move api logic to network layer
