+++ core/src/main/java/org/hisp/dhis/android/core/validation/internal/ValidationRuleService.kt
@@ -27,20 +27,18 @@
  */
 package org.hisp.dhis.android.core.validation.internal
 
-import org.hisp.dhis.android.core.arch.api.HttpServiceClient
-import org.hisp.dhis.android.core.arch.api.payload.internal.PayloadJackson
-import org.hisp.dhis.android.core.common.ObjectWithUid
 import org.hisp.dhis.android.core.validation.ValidationRule
+import org.hisp.dhis.android.network.common.HttpServiceClientKotlinx
 import org.hisp.dhis.android.network.common.fields.Fields
-import org.koin.core.annotation.Singleton
+import org.hisp.dhis.android.network.validationrule.ValidationRuleDatasetPayload
+import org.hisp.dhis.android.network.validationrule.ValidationRulePayload
 
-@Singleton
-internal class ValidationRuleService(private val client: HttpServiceClient) {
+internal class ValidationRuleService(private val client: HttpServiceClientKotlinx) {
     suspend fun getValidationRules(
         fields: Fields<ValidationRule>,
         uidsFilterString: String,
         paging: Boolean,
-    ): PayloadJackson<ValidationRule> {
+    ): ValidationRulePayload {
         return client.get {
             url("validationRules")
             parameters {
@@ -55,7 +53,7 @@ internal class ValidationRuleService(private val client: HttpServiceClient) {
         dataSetUid: String,
         id: String,
         paging: Boolean,
-    ): PayloadJackson<ObjectWithUid> {
+    ): ValidationRuleDatasetPayload {
         return client.get {
             url("validationRules")
             parameters {
